test controller user

+++ src/server/controllers/user/user.js
@@ -1,5 +1,5 @@
-import User from '../models/User';
-import Room from '../models/Room';
+const User = require('../../models/User');
+const Room = require('../../models/Room');
 
 const createUser = (username, socketId, cb) => {
   const newUser = new User({
@@ -10,7 +10,7 @@ const createUser = (username, socketId, cb) => {
     username,
   })
     .then(user => {
-      console.log(user)
+      // console.log(user)
       if (!user) {
         newUser.save()
         .then(user => {
@@ -18,8 +18,7 @@ const createUser = (username, socketId, cb) => {
           cb()
         })
         .catch(err => {
-          console.log(err)
-          cb('can\'t store the user in db')
+          return cb('can\'t store the user in db')
         })
       }
       else {
@@ -27,53 +26,51 @@ const createUser = (username, socketId, cb) => {
       }
     })
   .catch(err => {
-    console.log(err)
-    cb('can\'t store the user in db')
+    return cb('can\'t store the user in db')
   })
 }
 
 const deleteUser = (username, roomId, cb) => {
   User.deleteOne({ username })
   .then(user => {
-    console.log(`user ${user.username} removed from db`)
+    // console.log(user)
+    if (user.deletedCount ===  0) {
+      return cb('user doesn\'t exist in db')
+    }
+    // console.log(`user ${user.username} removed from db`)
     if (roomId !== null) {
       Room.findOne({ roomId })
       .then(room => {
-        console.log('testss', roomId)
         room.players = room.players.filter(u => u !== username)
         if (room.players.length === 0) {
-          Room.deleteOne({ roomId })
+          return Room.deleteOne({ roomId })
           .then(() => {
-            console.log(`room ${roomId} deleted because the user ${username} was the last player`)
+            // console.log(`room ${roomId} deleted because the user ${username} was the last player`)
             return cb()
           })
           .catch(err => {
-            console.log(err)
-            cb('something wrong happen while accesssing the db')
+            return cb('something wrong happen while accesssing the db')
           })
         }
-        room.save()
+        Room.updateOne({ _id: room._id }, { $set: { players: room.players } })
         .then(() => {
-          console.log(`the user ${username} has been deleted from the room ${roomId}`)
+          // console.log(`the user ${username} has been deleted from the room ${roomId}`)
           return cb()
         })
         .catch(err => {
-          console.log(err)
           cb('something wrong happen while accesssing the db')
         })
       })
       .catch(err => {
-        console.log(err)
         return cb('can\'t find the room in db')
       })
     }
     else {
-      console.log(`the room ${roomId} has not been found in db}`)
+      // console.log(`the room ${roomId} has not been found in db`)
       return cb('can not found room in db')
     }
   })
   .catch(err => {
-    console.log(err)
     cb('something wrong happen while accesssing the db')
   })
 }
